fear: se actualizan paginas y secciones. Ahora ademas se muestra el valor del saldo de SOL en USD.
