+++ src/app/sections/tokenslist-section.component.ts
@@ -2,11 +2,11 @@ import { Component, inject } from '@angular/core';
 import { toSignal } from '@angular/core/rxjs-interop';
 import { WalletStore } from '@heavy-duty/wallet-adapter';
 import { computedAsync } from 'ngxtension/computed-async';
-import { TokensList } from './shyft-api.services';
+import { TokensList } from '../shyft-api.services';
 import { MatCard } from '@angular/material/card';
 import { MatTableModule } from '@angular/material/table';
 import { MatDialog } from '@angular/material/dialog';
-import { TransferModalComponent } from './transfer-modal.component';
+import { TransferModalComponent } from '../transfer-modal.component';
 import { MatButton } from '@angular/material/button';
 import { DecimalPipe } from '@angular/common';
 
@@ -16,17 +16,21 @@ import { DecimalPipe } from '@angular/common';
     selector: 'billeterasol-tokenslist-section',
     imports: [MatTableModule, MatCard, MatButton, DecimalPipe],
     standalone: true,
-    styleUrls: ['../styles.scss'],
+    styleUrls: ['../../styles.scss'],
     template: `
     <mat-card class=" px-4 py-8 bg-black">
-      <h2 class="text-center justify text-3xl mb-4">Tokens</h2>
-
       @if (!allTokens()) {
         <p class="text-center">conect your wallet.</p>
       } @else if (allTokens()?.length === 0) {
         <p class="text-center">no Activity.</p>
       } @else  {
+        <div class="my-button-row center">
+        <button mat-raised-button color="primary" (click)="onTransfer()">SEND   </button>
+        <button mat-raised-button color="primary" (click)="onTransfer()">BUY    </button>
+        <button mat-raised-button color="primary" (click)="onTransfer()">RECEIVE</button>
+        </div>
         <table mat-table [dataSource]="allTokens() ?? []">
+          
           <ng-container matColumnDef="image">
             <th mat-header-cell *matHeaderCellDef></th>
             <td mat-cell *matCellDef="let element"><img [src]="element.info.image" [width]="50"/></td>
@@ -51,7 +55,7 @@ import { DecimalPipe } from '@angular/common';
           <tr mat-row *matRowDef="let row; columns: displayedColumns"></tr>
         </table>
         <footer class="flex justify-center px-4 py-8">
-                <button mat-raised-button color="primary" (click)="onTransfer()">Transfer tokens</button>
+                
         </footer>
       }
