+++ src/styles.scss
@@ -24,3 +24,17 @@ body {
   background-color: black;
 }
 
+.my-button-row button,
+ 
+.my-button-row a {
+  margin-right: 8px;
+  display: justify;
+  width: 100px;
+  align-items: center;
+}
+
+
+
+div.center {
+  text-align: center;
+}
